Cleaning up
cleaned up as much as I could right now, though there are still some left to do

+++ core/src/hu/nem3d/zincity/Misc/Builder.java
@@ -12,7 +12,7 @@ import java.util.ArrayList;
  * This handles the changes in a cell due to user input. Later in the code you can see which UiId and buildCode combinations refer to which action.
  */
 public class Builder {
-    private int UiId;
+    private int selectedMenuId;
     private int buildCode = 0;
     CityMap cityMap;
 
@@ -51,7 +51,7 @@ public class Builder {
      * @param city_
      */
     public Builder(int id, int code_, City city_){
-        this.UiId = id;
+        this.selectedMenuId = id;
         this.buildCode = code_;
         this.city = city_;
         cityMap = city.getCityMap();
@@ -64,25 +64,27 @@ public class Builder {
      * @param cell
      * @return
      */
-    public ArrayList<CityCell> build(CityCell cell){
+    public ArrayList<CityCell> build(CityCell cell) {
         ArrayList<CityCell> cells = new ArrayList<>();
         x = cell.getX();
         y = cell.getY();
-        switch (UiId){
-            case(1):
-                cells.add(buildZone(cell));
-                break;
-            case(2):
-                cells = buildSpecial(cell);
-                break;
-            case(5):
-                cells.add(buildNetwork(cell));
-                break;
-            case(6):
-                cells = deleteCell(cell);
-                break;
+        if (cell.getClass() != BlockedCell.class) {
+            switch (selectedMenuId) {
+                case (1):
+                    cells.add(buildZone(cell));
+                    break;
+                case (2):
+                    cells = buildSpecial(cell);
+                    break;
+                case (5):
+                    cells.add(buildNetwork(cell));
+                    break;
+                case (6):
+                    cells = deleteCell(cell);
+                    break;
+            }
+            cityMap.setBuildingLayer(buildLayer);
         }
-        cityMap.setBuildingLayer(buildLayer);
         return cells;
     }
 
@@ -96,28 +98,22 @@ public class Builder {
             switch (buildCode) {
                 case (1):
                     if(cell.getClass() == EmptyCell.class) {
-                        cell = new IndustrialZoneCell(2);
+                        cell = new IndustrialZoneCell(x,y,2);
                         cell.setTile(tileSet.getTile(8));
-                        cell.setX(x);
-                        cell.setY(y);
                         buildLayer.setCell(x, y, cell);
                     }
                     break;
                 case (2):
                     if(cell.getClass() == EmptyCell.class) {
-                        cell = new ServiceZoneCell(2);
+                        cell = new ServiceZoneCell(x,y,2);
                         cell.setTile(tileSet.getTile(11));
-                        cell.setX(x);
-                        cell.setY(y);
                         buildLayer.setCell(x, y, cell);
                     }
                     break;
                 case (3):
                     if(cell.getClass() == EmptyCell.class) {
-                        cell = new LivingZoneCell(4);
+                        cell = new LivingZoneCell(x,y,4);
                         cell.setTile(tileSet.getTile(5));
-                        cell.setX(x);
-                        cell.setY(y);
                         buildLayer.setCell(x, y, cell);
                     }
                     break;
@@ -138,8 +134,6 @@ public class Builder {
                             cell.setTile(tileSet.getTile(8 + ((IndustrialZoneCell) cell).getLevel() - 1));
                         }
                     }
-                    cell.setX(x);
-                    cell.setY(y);
                     buildLayer.setCell(x, y, cell);
                     break;
             }
@@ -161,11 +155,11 @@ public class Builder {
         if(cell.getClass() == EmptyCell.class) {
             switch (buildCode) {
                 case(1):
-                    cell = new PoliceCell(2,100);
+                    cell = new PoliceCell(x,y,2,100);
                     cell.setTile((tileSet.getTile(14)));
                     break;
                 case(2):
-                    cell = new FireStationCell(2,100);
+                    cell = new FireStationCell(x, y,2,100);
                     cell.setTile((tileSet.getTile(15)));
                     break;
                 case(3):
@@ -181,38 +175,18 @@ public class Builder {
                         }
                     }
                     if(isFree){
-                        cell = new ArenaCell(3, 100);
-                        ((ArenaCell) cell).setPart(BuildingCell.BuildingPart.valueOf("NorthWest"));
+                        int partArena = 0;
                         for(int i = 0; i > -2; i-- ){
                             for(int j = 0; j < 2; j++){
-                                ArenaCell tmpCell = new ArenaCell(3, 100);
-                                if(i == -1){
-                                    if(j == 1){
-                                        tmpCell.setPart(BuildingCell.BuildingPart.valueOf("SouthEast"));
-                                    }
-                                    else{
-                                        tmpCell.setPart(BuildingCell.BuildingPart.valueOf("SouthWest"));
-                                    }
-                                }
-                                else{
-                                    if(j == 1){
-                                        tmpCell.setPart(BuildingCell.BuildingPart.valueOf("NorthEast"));
-                                    }
-                                    else{
-                                        tmpCell.setPart(BuildingCell.BuildingPart.valueOf("NorthWest"));
-                                    }
-                                }
+                                BuildingCell.BuildingPart tmpPart = BuildingCell.BuildingPart.values()[partArena];
+                                partArena++;
+
+                                ArenaCell tmpCell = new ArenaCell(x+j,y+i,3, 100, tmpPart);
                                 tmpCell.setTile((tileSet.getTile(16 +tmpCell.getPart().ordinal())));
                                 buildLayer.setCell(x + j,y + i, tmpCell);
-                                tmpCell.setX(x+j);
-                                tmpCell.setY(y+i);
                                 returnCells.add(tmpCell);
                             }
                         }
-                        cell.setTile(tileSet.getTile(16));
-                        cell.setX(x);
-                        cell.setY(y);
-                        returnCells.add(cell);
                     }
                     break;
                 case(4):
@@ -228,38 +202,18 @@ public class Builder {
                         }
                     }
                     if(isFreeGen){
-                        cell = new GeneratorCell(3, 200);
-                        ((GeneratorCell) cell).setPart(BuildingCell.BuildingPart.valueOf("NorthWest"));;
+                        int partGen = 0;
                         for(int i = 0; i > -2; i-- ){
                             for(int j = 0; j < 2; j++){
-                                GeneratorCell tmpCell = new GeneratorCell(3, 200);
-                                if(i == -1){
-                                    if(j == 1){
-                                        tmpCell.setPart(BuildingCell.BuildingPart.valueOf("SouthEast"));
-                                    }
-                                    else{
-                                        tmpCell.setPart(BuildingCell.BuildingPart.valueOf("SouthWest"));
-                                    }
-                                }
-                                else{
-                                    if(j == 1){
-                                        tmpCell.setPart(BuildingCell.BuildingPart.valueOf("NorthEast"));
-                                    }
-                                    else{
-                                        tmpCell.setPart(BuildingCell.BuildingPart.valueOf("NorthWest"));
-                                    }
-                                }
-                                tmpCell.setTile((tileSet.getTile(20 + tmpCell.getPart().ordinal())));
+                                BuildingCell.BuildingPart tmpPart = BuildingCell.BuildingPart.values()[partGen];
+                                partGen++;
+
+                                GeneratorCell tmpCell = new GeneratorCell(x+j,y+i,3, 100, tmpPart);
+                                tmpCell.setTile((tileSet.getTile(20 +tmpCell.getPart().ordinal())));
                                 buildLayer.setCell(x + j,y + i, tmpCell);
-                                tmpCell.setX(x+j);
-                                tmpCell.setY(y+i);
                                 returnCells.add(tmpCell);
                             }
                         }
-                        cell.setTile(tileSet.getTile(20));
-                        cell.setX(x);
-                        cell.setY(y);
-                        returnCells.add(cell);
                     }
                     break;
 
@@ -270,8 +224,6 @@ public class Builder {
                     break;
             }
             if(buildCode != 3 && buildCode != 4){
-                cell.setX(x);
-                cell.setY(y);
                 buildLayer.setCell(x,y,cell);
                 returnCells.add(cell);
             }
@@ -290,10 +242,8 @@ public class Builder {
         switch (buildCode){
             case(2):
                 if(cell.getClass() == EmptyCell.class) {
-                    cell = new RoadCell();
+                    cell = new RoadCell(x,y);
                     cell.setTile((tileSet.getTile(4)));
-                    cell.setX(x);
-                    cell.setY(y);
                     buildLayer.setCell(x, y, cell);
                 }
                 break;
@@ -310,10 +260,12 @@ public class Builder {
      */
     public ArrayList <CityCell> deleteCell(CityCell cell){
         ArrayList <CityCell> returnCells = new ArrayList<>();
-        if(cell.getClass() != BlockedCell.class){   //TODO implement conflicting delete here in if --> example: cell.ConflictDelete in if statement
-            if(cell.getClass() == LivingZoneCell.class){
+        //This is currently the best way to solve this since the Zones do not know who is in them and I have to check what kind of zone it is
+        //So when I get the type of cell I have to iterate through all the citizens and check for everyone's  home and workplace, cant do much else if I
+        //Understand the structure correctly
+         if(cell.getClass() == LivingZoneCell.class){         //TODO implement conflicting delete here in if --> example: cell.ConflictDelete in if statement
                 for (Citizen citizen : city.citizens) {
-                    if(citizen.getHome().getX() == y && citizen.getHome().getY() == x){
+                    if(citizen.getHome().getX() == x && citizen.getHome().getY() == y){
                         citizen.getWorkplace().removeOccupant(citizen);
                         citizen.setWorkplace(null);
                         citizen.setHome(null);
@@ -323,43 +275,42 @@ public class Builder {
             }
             else if(cell.getClass() == ServiceZoneCell.class || cell.getClass() == IndustrialZoneCell.class){
                 for (Citizen citizen : city.citizens) {
-                    if(citizen.getWorkplace().getX() == y && citizen.getWorkplace().getY() == x){
+                    if(citizen.getWorkplace().getX() == x && citizen.getWorkplace().getY() == y){
                         citizen.getWorkplace().removeOccupant(citizen);
                         citizen.setWorkplace(null);
                     }
                 }
             }
+
+
             if(cell.getClass() == ArenaCell.class || cell.getClass() == GeneratorCell.class){
-                BuildingCell.BuildingPart part = ((BuildingCell) cell).getPart();
-                if(part == BuildingCell.BuildingPart.valueOf("NorthEast")){
-                    x = x - 1;
-                }
-                else if(part == BuildingCell.BuildingPart.valueOf("SouthWest")){
-                    y = y + 1;
-                }
-                else if(part == BuildingCell.BuildingPart.valueOf("SouthEast")){
-                    x = x - 1;
-                    y = y + 1;
+                int partValue = ((BuildingCell) cell).getPart().ordinal();
+                switch (partValue){
+                    case(1):
+                        x = x - 1;
+                        break;
+                    case(2):
+                        y = y + 1;
+                        break;
+                    case(3):
+                        x = x - 1;
+                        y = y + 1;
+                        break;
                 }
 
                 for(int i = 0; i > -2; i-- ){
                     for(int j = 0; j < 2; j++){
-                        EmptyCell tmpCell = new EmptyCell();
+                        EmptyCell tmpCell = new EmptyCell(x+j,y+i);
                         tmpCell.setTile((tileSet.getTile(0)));
                         buildLayer.setCell(x + j,y + i, tmpCell);
-                        tmpCell.setX(x+j);
-                        tmpCell.setY(y+i);
                         returnCells.add(tmpCell);
                     }
                 }
 
             }
-            cell = new EmptyCell();
+            cell = new EmptyCell(x,y);
             cell.setTile((tileSet.getTile(0)));
-            cell.setX(x);
-            cell.setY(y);
             buildLayer.setCell(x, y, cell);
-        }
         return  returnCells;
     }
